Remove unnecessary symbol visibility to eliminate conflict with postgres_fdw
Symbol deparseSelectSql() is used in both postgres_fdw and tds_fdw. This
symbol is exported as a global symbol in both cases as is the default
behavior for non-static functions. When tds_fdw.so is loaded before
postgres_fdw.so linker will link the implementation from tds_fdw into
postgres_fdw, causing segfaults when postgres_fdw is used.

Because there is no reason to actually export these functions this patch
adjusts build flags to default functions to non-exported and overrides
PostgreSQL PGDLLEXPORT definition to mark functions as exported. As
PG_FUNCTION_INFO_V1 does not add export flag to the actual function
declaration we need to manually annotate the exported functions.

An alternative solution would be to just rename the conflicting functions.
For platforms without -fvisibility=hidden support that might be a good
idea anyway.