Remove bad pfree of query string
tdsEndForeignScan tries to free the query string, but is just a
pointer into the ForeignScanState we got passed in tdsBeginForeignScan.
We have no business to free that memory, which is allocated in
PortalContext by the traffic cop and freed when the portal is done.

This leads to intermittent crashes when the pfree segfaults.